Get from 0 to a basic site with JSP handling
This is BEFORE Expression Langage comes into play.
PLEASE note that the default Servlet has been redefined : this has led
to a lot of confusion, tests (and as always learning!) these last days.
Indeed, I was mixing this OCR course with a pre-existing "basic" J2EE
project, and these two were not COMPLETLY aligned.

I noticed that when using .include("static ressources") it seems to fallback,
eventually, on the default Servlet. But this (redefined)servlet were only
displaying a single static page, thus leading to the site being frozen, not-responsive.

TAKEAWAY: just debug! A few breakpoints here and there, and boum, I understood the issue!
(that I was [embarrassingly] investigating for.. a  time too long to tell!)
The mapping DID work (/login & /logout) but .include() redirects to default servlet,
which here broke the proper flow and interrupted prematuraly.